Workaround bug with missing operatingsystem and osfamily facts
See bug https://tickets.puppetlabs.com/browse/PUP-8543

When there is no operatingsystem or osfamily fact available on the host
os, tests will fail. For most tests, this is not a problem since they
are looped inside on_supported_os to supply their initial facts.
Functions to not use this which leads to puppet failing to load.